+++ src/addons/display_commander/config/default_overrides.cpp
@@ -72,7 +72,6 @@ static const KeyDisplayName k_key_display_names[] = {
     {"ContinueRendering", "Continue Rendering in Background (Fake Fullscreen)"},
     {"PreventMinimize", "Prevent Minimize"},
     {"PreventAlwaysOnTop", "Prevent Always on Top"},
-    {"EnableFlipChain", "Enable Flip Chain"},
     {"AutoColorspace", "Auto color space"},
     {"WindowMode", "Window Mode"},
 };